major fix

+++ enma_pe/pe_bound_import.h
@@ -2,27 +2,27 @@
 
 class bound_imported_ref{
     std::string ref_name;
-    DWORD timestamp;
+    uint32_t  timestamp;
     
 public:
     bound_imported_ref::bound_imported_ref();
     bound_imported_ref::bound_imported_ref(const bound_imported_ref& ref);
-    bound_imported_ref::bound_imported_ref(const std::string& ref_name, DWORD timestamp);
+    bound_imported_ref::bound_imported_ref(const std::string& ref_name, uint32_t  timestamp);
     bound_imported_ref::~bound_imported_ref();
 
     bound_imported_ref& bound_imported_ref::operator=(const bound_imported_ref& ref);
 public:
     void bound_imported_ref::set_ref_name(const std::string& ref_name);
-    void bound_imported_ref::set_timestamp(DWORD timestamp);
+    void bound_imported_ref::set_timestamp(uint32_t  timestamp);
 
 public:
     std::string bound_imported_ref::get_ref_name() const;
-    DWORD bound_imported_ref::get_timestamp() const;
+    uint32_t  bound_imported_ref::get_timestamp() const;
 };
 
 class bound_imported_library {
     std::string library_name;
-    DWORD timestamp;
+    uint32_t  timestamp;
     std::vector<bound_imported_ref> refs;
 public:
     bound_imported_library::bound_imported_library();
@@ -32,13 +32,13 @@ public:
     bound_imported_library& bound_imported_library::operator=(const bound_imported_library& lib);
 public:
     void bound_imported_library::set_library_name(const std::string& library_name);
-    void bound_imported_library::set_timestamp(DWORD timestamp);
+    void bound_imported_library::set_timestamp(uint32_t  timestamp);
 
     void bound_imported_library::add_ref(const bound_imported_ref& ref);
 public:
     std::string bound_imported_library::get_library_name() const;
-    DWORD bound_imported_library::get_timestamp() const;
-    unsigned int bound_imported_library::get_number_of_forwarder_refs() const;
+    uint32_t  bound_imported_library::get_timestamp() const;
+    size_t bound_imported_library::get_number_of_forwarder_refs() const;
 
     std::vector<bound_imported_ref>& bound_imported_library::get_refs();
 };
