+++ enma_pe/pe_relocations.h
@@ -1,9 +1,9 @@
 #pragma once
 
 struct relocation_item {
-	DWORD relative_virtual_address;
-	DWORD relocation_id;
-    DWORD64 data;
+    uint32_t relative_virtual_address;
+    uint32_t relocation_id;
+    uint64_t data;
 };
 
 
@@ -16,20 +16,20 @@ public:
 
     relocation_table& relocation_table::operator=(const relocation_table& relocations);
 public:
-    void relocation_table::add_item(DWORD rva, DWORD relocation_id);
-    bool relocation_table::erase_item(DWORD rva);
-    bool relocation_table::erase_first_item_by_id(DWORD relocation_id);
-    bool relocation_table::erase_all_items_by_id(DWORD relocation_id);
+    void relocation_table::add_item(uint32_t rva, uint32_t relocation_id);
+    bool relocation_table::erase_item(uint32_t rva);
+    bool relocation_table::erase_first_item_by_id(uint32_t relocation_id);
+    bool relocation_table::erase_all_items_by_id(uint32_t relocation_id);
 
-    void relocation_table::get_items_by_relocation_id(std::vector<relocation_item*>& found_relocs,DWORD relocation_id);
+    void relocation_table::get_items_by_relocation_id(std::vector<relocation_item*>& found_relocs, uint32_t relocation_id);
 
     void relocation_table::clear();
     void relocation_table::sort();  
 public:
-    unsigned int relocation_table::size() const;
-    bool relocation_table::has_item(DWORD rva) const;
-    bool relocation_table::has_item_id(DWORD relocation_id) const;
-    bool relocation_table::get_item_id(DWORD rva, DWORD& relocation_id) const;
+    size_t relocation_table::size() const;
+    bool relocation_table::has_item(uint32_t rva) const;
+    bool relocation_table::has_item_id(uint32_t relocation_id) const;
+    bool relocation_table::get_item_id(uint32_t rva, uint32_t& relocation_id) const;
 
     std::vector<relocation_item>& relocation_table::get_items();
 };
